supporting auth wizard URL state, refactor wizard steps
