+++ package.json
@@ -9,6 +9,8 @@
     "dev:app": "turbo run @propsto/app#dev",
     "dev:auth": "turbo run @propsto/auth#dev @propsto/email#dev @propsto/data#db-studio",
     "dev:email": "turbo run @propsto/email#dev",
+    "dev:nuke": "turbo run @propsto/data#db-nuke",
+    "dev:setup": "turbo run @propsto/data#db-setup",
     "dev:web": "turbo run @propsto/web#dev",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
     "health": "turbo run lint type-check",
@@ -16,9 +18,7 @@
     "lint": "turbo run lint --force",
     "prepare": "husky",
     "test": "turbo run test --force",
-    "type-check": "turbo run type-check",
-    "dev:setup": "turbo run @propsto/data#db-setup",
-    "dev:nuke": "turbo run @propsto/data#db-nuke"
+    "type-check": "turbo run type-check"
   },
   "dependencies": {
     "@upstash/ratelimit": "2.0.5",
@@ -26,34 +26,35 @@
     "@vercel/kv": "3.0.0",
     "autoprefixer": "10.4.20",
     "cal-sans": "1.0.1",
-    "lucide-react": "0.468.0",
-    "next": "15.1.1",
+    "lucide-react": "0.469.0",
+    "next": "15.1.2",
     "next-themes": "0.4.4",
     "nodemailer": "6.9.16",
     "postcss": "8.4.49",
     "react": "19.0.0",
     "react-dom": "19.0.0",
-    "react-hook-form": "7.54.0",
+    "react-hook-form": "7.54.2",
     "tailwind-merge": "2.5.5",
-    "tailwindcss": "3.4.16",
+    "tailwindcss": "3.4.17",
     "tailwindcss-animate": "1.0.7",
     "typescript": "5.7.2",
     "vitest": "2.1.8",
-    "zod": "3.23.8"
+    "zod": "3.24.1"
   },
   "devDependencies": {
-    "@playwright/test": "1.49.0",
-    "@trivago/prettier-plugin-sort-imports": "5.1.0",
-    "@types/node": "20.14.2",
+    "@next/eslint-plugin-next": "15.1.2",
+    "@playwright/test": "1.49.1",
+    "@trivago/prettier-plugin-sort-imports": "5.2.0",
+    "@types/node": "22.10.2",
     "@types/nodemailer": "6.4.17",
-    "@types/react": "19.0.1",
-    "@types/react-dom": "19.0.1",
-    "@typescript-eslint/typescript-estree": "8.17.0",
-    "dotenv-cli": "7.4.4",
+    "@types/react": "19.0.2",
+    "@types/react-dom": "19.0.2",
+    "@typescript-eslint/typescript-estree": "8.18.1",
+    "dotenv-cli": "8.0.0",
     "eslint": "8.57.0",
     "framer-motion": "v12.0.0-alpha.2",
     "husky": "9.1.7",
-    "npm-check-updates": "^17.1.11",
+    "npm-check-updates": "17.1.12",
     "prettier": "3.4.2",
     "sort-package-json": "^2.12.0",
     "tsx": "4.19.2",
@@ -62,17 +63,18 @@
   "peerDependencies": {
     "@propsto/config": "workspace:*"
   },
-  "packageManager": "pnpm@9.15.0",
+  "packageManager": "pnpm@9.15.1",
   "pnpm": {
     "overrides": {
-      "@types/react": "19.0.1",
-      "@types/react-dom": "19.0.1",
-      "@typescript-eslint/typescript-estree": "8.17.0",
-      "eslint": "8.57.0",
-      "next": "15.1.1",
-      "react": "19.0.0",
-      "react-dom": "19.0.0",
-      "typescript": "5.7.2"
+      "@types/react": "$@types/react",
+      "@types/react-dom": "$@types/react-dom",
+      "@typescript-eslint/typescript-estree": "$@typescript-eslint/typescript-estree",
+      "@next/eslint-plugin-next": "$@next/eslint-plugin-next",
+      "eslint": "$eslint",
+      "next": "$next",
+      "react": "$react",
+      "react-dom": "$react-dom",
+      "typescript": "$typescript"
     }
   }
 }
