+++ packages/config/package.json
@@ -8,8 +8,10 @@
     "eslint-plugin-local-rules": "3.0.2"
   },
   "devDependencies": {
-    "@next/eslint-plugin-next": "14.2.4",
     "@vercel/style-guide": "6.0.0",
-    "eslint-config-turbo": "2.0.4"
+    "eslint-config-turbo": "2.3.3"
+  },
+  "peerDependencies": {
+    "@next/eslint-plugin-next": "*"
   }
 }
