+++ packages/email/package.json
@@ -11,12 +11,12 @@
     "start": "email dev --dir templates"
   },
   "dependencies": {
-    "@react-email/components": "0.0.30",
-    "react-email": "3.0.3",
+    "@react-email/components": "0.0.31",
+    "react-email": "3.0.4",
     "resend": "4.0.1"
   },
   "devDependencies": {
-    "maildev": "2.1.0"
+    "maildev": "2.2.1"
   },
   "peerDependencies": {
     "@propsto/config": "workspace:*",
