+++ packages/ui/package.json
@@ -7,14 +7,14 @@
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "@radix-ui/react-avatar": "1.1.1",
-    "@radix-ui/react-collapsible": "1.1.1",
-    "@radix-ui/react-dialog": "1.1.2",
-    "@radix-ui/react-dropdown-menu": "2.1.2",
-    "@radix-ui/react-label": "2.1.0",
-    "@radix-ui/react-navigation-menu": "1.2.1",
-    "@radix-ui/react-separator": "1.1.0",
-    "@radix-ui/react-slot": "1.1.0",
+    "@radix-ui/react-avatar": "1.1.2",
+    "@radix-ui/react-collapsible": "1.1.2",
+    "@radix-ui/react-dialog": "1.1.4",
+    "@radix-ui/react-dropdown-menu": "2.1.4",
+    "@radix-ui/react-label": "2.1.1",
+    "@radix-ui/react-navigation-menu": "1.2.3",
+    "@radix-ui/react-separator": "1.1.1",
+    "@radix-ui/react-slot": "1.1.1",
     "@tailwindcss/forms": "0.5.9",
     "class-variance-authority": "0.7.1",
     "clsx": "2.1.1"
